move authenticator model and tests
