+++ src/sentry/api/endpoints/auth_index.py
@@ -21,8 +21,8 @@ from sentry.auth.authenticators.u2f import U2fInterface
 from sentry.auth.providers.saml2.provider import handle_saml_single_logout
 from sentry.auth.services.auth.impl import promote_request_rpc_user
 from sentry.auth.superuser import SUPERUSER_ORG_ID
-from sentry.models.authenticator import Authenticator
 from sentry.organizations.services.organization import organization_service
+from sentry.users.models.authenticator import Authenticator
 from sentry.utils import auth, json, metrics
 from sentry.utils.auth import DISABLE_SSO_CHECK_FOR_LOCAL_DEV, has_completed_sso, initiate_login
 from sentry.utils.settings import is_self_hosted
