+++ src/sentry/api/serializers/models/user.py
@@ -17,7 +17,6 @@ from sentry.api.serializers.types import SerializedAvatarFields
 from sentry.app import env
 from sentry.auth.elevated_mode import has_elevated_mode
 from sentry.hybridcloud.services.organization_mapping import organization_mapping_service
-from sentry.models.authenticator import Authenticator
 from sentry.models.authidentity import AuthIdentity
 from sentry.models.avatars.user_avatar import UserAvatar
 from sentry.models.options.user_option import UserOption
@@ -29,6 +28,7 @@ from sentry.models.useremail import UserEmail
 from sentry.models.userpermission import UserPermission
 from sentry.models.userrole import UserRoleUser
 from sentry.organizations.services.organization import RpcOrganizationSummary
+from sentry.users.models.authenticator import Authenticator
 from sentry.users.services.user import RpcUser
 from sentry.utils.avatar import get_gravatar_url
