+++ src/sentry/auth/authenticators/base.py
@@ -14,8 +14,8 @@ from sentry.utils.otp import TOTP, generate_secret_key
 if TYPE_CHECKING:
     from django.utils.functional import _StrPromise
 
-    from sentry.models.authenticator import Authenticator
     from sentry.models.user import User
+    from sentry.users.models.authenticator import Authenticator
 
 
 class ActivationResult:
@@ -150,7 +150,7 @@ class AuthenticatorInterface:
 
         If `disallow_new_enrollment` is `True`, raises exception: `NewEnrollmentDisallowed`.
         """
-        from sentry.models.authenticator import Authenticator
+        from sentry.users.models.authenticator import Authenticator
 
         if self.disallow_new_enrollment:
             raise NewEnrollmentDisallowed
