+++ src/sentry/models/user.py
@@ -38,7 +38,6 @@ from sentry.db.postgres.transactions import enforce_constraints
 from sentry.hybridcloud.outbox.category import OutboxCategory
 from sentry.integrations.types import EXTERNAL_PROVIDERS, ExternalProviders
 from sentry.locks import locks
-from sentry.models.authenticator import Authenticator
 from sentry.models.avatars import UserAvatar
 from sentry.models.lostpasswordhash import LostPasswordHash
 from sentry.models.organizationmapping import OrganizationMapping
@@ -47,6 +46,7 @@ from sentry.models.orgauthtoken import OrgAuthToken
 from sentry.models.outbox import ControlOutboxBase, outbox_context
 from sentry.organizations.services.organization import RpcRegionUser, organization_service
 from sentry.types.region import find_all_region_names, find_regions_for_user
+from sentry.users.models.authenticator import Authenticator
 from sentry.users.services.user import RpcUser
 from sentry.utils.http import absolute_uri
 from sentry.utils.retries import TimedRetryPolicy
@@ -333,13 +333,13 @@ class User(Model, AbstractBaseUser):
     def merge_to(from_user: User, to_user: User) -> None:
         # TODO: we could discover relations automatically and make this useful
         from sentry.models.auditlogentry import AuditLogEntry
-        from sentry.models.authenticator import Authenticator
         from sentry.models.authidentity import AuthIdentity
         from sentry.models.avatars.user_avatar import UserAvatar
         from sentry.models.identity import Identity
         from sentry.models.options.user_option import UserOption
         from sentry.models.organizationmembermapping import OrganizationMemberMapping
         from sentry.models.useremail import UserEmail
+        from sentry.users.models.authenticator import Authenticator
 
         from_user_id = from_user.id
         to_user_id = to_user.id
