+++ src/sentry/web/frontend/debug/debug_mfa_removed_email.py
@@ -3,8 +3,8 @@ import datetime
 from django.http import HttpRequest, HttpResponse
 from django.views.generic import View
 
-from sentry.models.authenticator import Authenticator
 from sentry.security.emails import generate_security_email
+from sentry.users.models.authenticator import Authenticator
 
 from .mail import MailPreview
