+++ tests/sentry/api/endpoints/test_user_authenticator_enroll.py
@@ -7,7 +7,6 @@ from django.urls import reverse
 
 from sentry import audit_log
 from sentry.models.auditlogentry import AuditLogEntry
-from sentry.models.authenticator import Authenticator
 from sentry.models.organization import Organization
 from sentry.models.organizationmember import OrganizationMember
 from sentry.models.useremail import UserEmail
@@ -18,6 +17,7 @@ from sentry.testutils.cases import APITestCase
 from sentry.testutils.helpers import override_options
 from sentry.testutils.outbox import outbox_runner
 from sentry.testutils.silo import assume_test_silo_mode, control_silo_test
+from sentry.users.models.authenticator import Authenticator
 from tests.sentry.api.endpoints.test_user_authenticator_details import assert_security_email_sent
 
 
@@ -479,8 +479,9 @@ class AcceptOrganizationInviteTest(APITestCase):
 
         # Mutate the OrganizationMember, putting it out of sync with the
         # pending member cookie.
-        with assume_test_silo_mode(SiloMode.REGION), unguarded_write(
-            using=router.db_for_write(OrganizationMember)
+        with (
+            assume_test_silo_mode(SiloMode.REGION),
+            unguarded_write(using=router.db_for_write(OrganizationMember)),
         ):
             om.update(id=om.id + 1)
 
@@ -501,8 +502,9 @@ class AcceptOrganizationInviteTest(APITestCase):
 
         # Mutate the OrganizationMember, putting it out of sync with the
         # pending member cookie.
-        with assume_test_silo_mode(SiloMode.REGION), unguarded_write(
-            using=router.db_for_write(OrganizationMember)
+        with (
+            assume_test_silo_mode(SiloMode.REGION),
+            unguarded_write(using=router.db_for_write(OrganizationMember)),
         ):
             om.update(token="123")
