+++ tests/sentry/api/serializers/test_user.py
@@ -1,7 +1,6 @@
 from sentry.api.serializers import serialize
 from sentry.api.serializers.models.user import DetailedSelfUserSerializer, DetailedUserSerializer
 from sentry.auth.authenticators import available_authenticators
-from sentry.models.authenticator import Authenticator
 from sentry.models.authidentity import AuthIdentity
 from sentry.models.authprovider import AuthProvider
 from sentry.models.avatars.user_avatar import UserAvatar
@@ -9,6 +8,7 @@ from sentry.models.useremail import UserEmail
 from sentry.models.userpermission import UserPermission
 from sentry.testutils.cases import TestCase
 from sentry.testutils.silo import control_silo_test
+from sentry.users.models.authenticator import Authenticator
 
 
 @control_silo_test
