+++ tests/tools/test.txt
@@ -1,0 +1,21 @@
+tests/sentry/rules/processing/test_delayed_processing.py
+tests/sentry/search/test_utils.py
+tests/sentry/silo/test_util.py
+tests/sentry/snuba/test_entity_subscriptions.py
+tests/sentry/snuba/test_errors.py
+tests/sentry/snuba/test_profiles.py
+tests/sentry/snuba/test_subscriptions.py
+tests/sentry/sudo/test_middleware.py
+tests/sentry/sudo/test_utils.py
+tests/sentry/sudo/test_views.py
+tests/sentry/tasks/deletion/test_hybrid_cloud.py
+tests/sentry/tasks/test_auto_remove_inbox.py
+tests/sentry/tasks/test_backfill_outboxes.py
+
+
+tests/sentry/rules/processing/test_delayed_processing.py tests/sentry/search/test_utils.py tests/sentry/silo/test_util.py tests/sentry/snuba/test_entity_subscriptions.py
+tests/sentry/snuba/test_errors.py tests/sentry/snuba/test_profiles.py tests/sentry/snuba/test_subscriptions.py tests/sentry/sudo/test_middleware.py tests/sentry/sudo/test_utils.py
+
+use
+tests/sentry/sudo/test_views.py tests/sentry/tasks/deletion/test_hybrid_cloud.py tests/sentry/tasks/test_auto_remove_inbox.py tests/sentry/tasks/test_backfill_outboxes.py
+tests/sentry/tasks/deletion/test_hybrid_cloud.py tests/sentry/tasks/test_auto_remove_inbox.py tests/sentry/tasks/test_backfill_outboxes.py
